cleaning code  and add Comments
