+++ src/js/counter.js
@@ -1,16 +1,10 @@
 import { todoList } from ".."
 export function counter() {
   //localStorage
-  // let localArr = JSON.parse(localStorage.getItem('mydayapp-js'))
-  // let completedArr = JSON.parse(localStorage.getItem('mydayapp-js-completed'))
-  // let pendingArr = JSON.parse(localStorage.getItem('mydayapp-js-pending'))
-
   const span = document.querySelector(".todo-count")
-  const strong = document.querySelector("strong")
   const numberOfItem = todoList.filter(element => element.visible === true
   );
-  console.log("lengthList", numberOfItem.length);
-
+  //renderización de la cantidad de tareas  
   if (numberOfItem.length != 1) {
     span.innerHTML = ""
     span.innerHTML = `<strong>${numberOfItem.length}</strong> items left`;
