+++ src/js/navegationBtn.js
@@ -3,12 +3,12 @@ import { newItem } from "./newItem";
 import { counter } from "./counter";
 
 export function navegation() {
+  //Obteniendo elementos del DOM
   const allBtn = document.querySelector(".allBtn");
   const completedBtn = document.querySelector(".completedBtn");
   const pendingBtn = document.querySelector(".pendingBtn");
-
-  if (location.hash.startsWith('#/completed')) {
-    //const completedArrGet = JSON.parse(localStorage.getItem('mydayapp-js-completed'));
+  //Verificación del hash para filtar el contenido del render: all, pending y completed
+  if (location.hash === '#/completed') {
     todoList.forEach(element => {
       if (element.completed === false) {
         element.visible = false
@@ -20,10 +20,9 @@ export function navegation() {
     allBtn.classList.remove("selected");
     completedBtn.classList.add("selected");
     pendingBtn.classList.remove("selected");
-    newItem(todoList);
+    newItem();
     counter()
-  } else if (location.hash.startsWith('#/pending')) {
-    //const pendingArr = JSON.parse(localStorage.getItem('mydayapp-js-pending'));
+  } else if (location.hash === '#/pending') {
     todoList.forEach(element => {
       if (element.completed === true) {
         element.visible = false
@@ -35,10 +34,10 @@ export function navegation() {
     allBtn.classList.remove("selected");
     completedBtn.classList.remove("selected");
     pendingBtn.classList.add("selected");
-    newItem(todoList)
+    newItem()
     counter()
 
-  } else if (location.hash.startsWith('#/all')) {
+  } else if (location.hash === '#/all') {
     todoList.forEach(element => {
       element.visible = true
 
@@ -46,7 +45,7 @@ export function navegation() {
     allBtn.classList.add("selected");
     completedBtn.classList.remove("selected");
     pendingBtn.classList.remove("selected");
-    newItem(todoList)
+    newItem()
     counter()
   }
