+++ src/js/newItem.js
@@ -1,25 +1,15 @@
 import { counter } from "./counter";
 import { todoList } from "..";
-import { btnClearCompletedVisible } from "..";
+import { btnClearCompletedVisible } from "./btnClearCompletedVisible";
 
 export function newItem() {
-  //LocalStorage
-  // if (location.hash.startsWith('#/')) {
-  // }
-
-  //const todoList = JSON.parse(localStorage.getItem("mydayapp-js"));
-
-  //pendiente
-  // let itemCompleted = todoList.filter(item => item.completed);
-  // console.log("itemCompleted", itemCompleted);
-
+  // verificación de tareas pendientes
   btnClearCompletedVisible()
-
+  // Limpieza de los contenedores
   const ul = document.querySelector(".todo-list");
   ul.innerHTML = "";
   let newTodo = document.querySelector(".new-todo");
   newTodo.value = "";
-  // const newText = text.trim();
 
   //ocular main y footer hasta agregar un archivo
   const main = document.querySelector(".main");
@@ -32,10 +22,8 @@ export function newItem() {
     main.classList.add("hidden");
     footer.classList.add("hidden");
   }
-
+  //iteración por cada elemento para hacer render
   todoList.forEach(item => {
-    //console.log("todoListitemmmm", item);
-
     //creación de elementos HTML
     const li = document.createElement("li");
     const div = document.createElement("div");
@@ -48,33 +36,10 @@ export function newItem() {
     input2.classList.add("edit");
     input2.value = item.tarea;
 
-    input2.addEventListener("keydown", (e) => {
-      if (e.code === "Enter") {
-        //input2.value = input2.value;
-        item.tarea = input2.value.trim()
-        input2.parentNode.classList.remove('editing');
-        label.innerHTML = "";
-        label.append(input2.value.trim())
-
-        localStorage.setItem('mydayapp-js', JSON.stringify(todoList))
-      } else if (e.code === "Escape") {
-        input2.value = item.tarea.trim();
-        input2.parentNode.classList.remove('editing');
-      }
-    });
-
-    //validación de contenido vacio
-    // if (newText.length != 0) {
-
     //agregando atributos a los inputs
     div.classList.add("view");
     input.classList.add("toggle");
     input.setAttribute("type", "checkbox");
-
-    // if (isCheked === true) {
-    //   li.classList.add("completed");
-    //   input.checked = true;
-    // }
 
     if (item.completed === true) {
       input.checked = true
@@ -87,29 +52,27 @@ export function newItem() {
       div.classList.remove("hidden");
     }
 
-
-    //Escuchar eventos  de doble click y cambio de estado en el checkbox
+    // toggle chackbox
     input.addEventListener('click', () => {
       if (input.checked) {
         li.classList.add("completed");
         item.completed = true;
         localStorage.setItem('mydayapp-js', JSON.stringify(todoList))
         btnClearCompletedVisible()
-        console.log('input', item.completed);
       } else {
         li.classList.remove("completed");
         item.completed = false;
         localStorage.setItem('mydayapp-js', JSON.stringify(todoList))
-        console.log('input2', item.completed);
         btnClearCompletedVisible()
 
       }
     })
+    //Escuchar eventos  de doble click y cambio de estado 
     label.ondblclick = function () {
       if (this.checked) {
         li.classList.remove("completed");
         li.classList.add("editing");
-        input2.focus()
+        input2.focus().trim()
       } else if (!this.checked) {
         li.classList.add("editing");
         input2.focus()
@@ -119,36 +82,38 @@ export function newItem() {
       }
 
     }
+    // Escuchador de eventos para guardar y renderizar los cambios realizardos
+    input2.addEventListener("keydown", (e) => {
+      if (e.code === "Enter") {
+        item.tarea = input2.value
+        input2.parentNode.classList.remove('editing');
+        label.innerHTML = "";
+        label.append(input2.value)
+        localStorage.setItem('mydayapp-js', JSON.stringify(todoList))
+      } else if (e.code === "Escape") {
+        input2.value = item.tarea;
+        input2.parentNode.classList.remove('editing');
+      }
+    });
 
+    //eliminación de todo en el Btn destroy
     btn.classList.add("destroy");
     btn.addEventListener('click', () => {
       const findIndex = todoList.findIndex(element => element === item);
       todoList.splice(findIndex, 1)
       localStorage.setItem('mydayapp-js', JSON.stringify(todoList))
-      newItem(todoList)
-      console.log("itemDestroy", findIndex);
+      newItem()
     })
 
     //isercion de los elementos creados en el HTML
     ul.appendChild(li);
     li.appendChild(div);
     li.appendChild(input2);
-    label.append(input2.value.trim());
+    label.append(input2.value);
     div.appendChild(input);
     div.appendChild(label);
     div.appendChild(btn);
-    // } else {
-    //   alert("Write something")
-    // }
-    //console.log(todoList);
   })
 
-
-
-
-
-
-
-
-  counter(0)
+  counter()
 }
